Refactor MCP client
